requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
plotly>=5.19.0
python-dotenv>=1.0.0
pandas>=2.0.0
duckduckgo-search>=6.0.0
groq>=0.9.0
